refactor(core): split auth helpers
